# Use compiled expression in TypedClientBuilder