This should shave off some milliseconds every time a typed client is constructed :smile:

There's already tests covering these paths, but I added an extra benchmark, shown below.

## Before

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10.0.17134
Intel Core i7-6820HQ CPU 2.70GHz (Skylake), 1 CPU, 8 logical cores and 4 physical cores
.NET Core SDK=5.0.100-alpha1-013788
  [Host]     : .NET Core 5.0.0-alpha1.19404.5 (CoreCLR 5.0.19.40302, CoreFX 5.0.19.38102), 64bit RyuJIT
  Job-FBDBIV : .NET Core 5.0.0-alpha1.19404.5 (CoreCLR 5.0.19.40302, CoreFX 5.0.19.38102), 64bit RyuJIT

Runtime=Core  Server=True  Toolchain=.NET 5.0  
RunStrategy=Throughput  

```
| Method |     Mean |    Error |   StdDev |        Op/s |  Gen 0 | Allocated |
|------- |---------:|---------:|---------:|------------:|-------:|----------:|
|  Build | 692.2 ns | 15.00 ns | 17.27 ns | 1,444,691.6 | 0.0095 |     408 B |

## After

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10.0.17134
Intel Core i7-6820HQ CPU 2.70GHz (Skylake), 1 CPU, 8 logical cores and 4 physical cores
.NET Core SDK=5.0.100-alpha1-013788
  [Host]     : .NET Core 5.0.0-alpha1.19404.5 (CoreCLR 5.0.19.40302, CoreFX 5.0.19.38102), 64bit RyuJIT
  Job-TKTSZD : .NET Core 5.0.0-alpha1.19404.5 (CoreCLR 5.0.19.40302, CoreFX 5.0.19.38102), 64bit RyuJIT

Runtime=Core  Server=True  Toolchain=.NET 5.0  
RunStrategy=Throughput  

```
| Method |     Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|------- |---------:|----------:|----------:|-------------:|-------:|----------:|
|  Build | 14.88 ns | 0.3769 ns | 0.4032 ns | 67,220,764.5 | 0.0006 |      24 B |