@BrennanConroy 2019-10-01
Maybe I'm missing it, but you seem to be missing some recent changes to this file, could you rebase on master?