@khellang 2019-10-01
> Maybe I'm missing it, but you seem to be missing some recent changes to this file, could you rebase on master?

Hmm, looks like I based it of my origin master which is pretty far behind, but GitHub says it's automatically mergeable though...