@khellang 2019-10-02
Alright, got rid of `System.Linq.Expressions` and rebased 👌