@khellang 2019-10-02
Hmm, I think it might be even cleaner to generate a static `Build` method on the actual `Impl` type and create a delegate directly from that.