@davidfowl 2019-10-08
/azp run AspNetCore-ci